Keep a modulator's amount fractional, and compute it in double
The GS controller matrix sets a modulator's amount from a normalized
controller value: amplitude control at 127 asks for 984.375, not 984.
Upstream assigns that straight to transformAmount, which is an ordinary
number; the port's field was an int16_t and the call sites cast to it,
so every matrix entry landed on a whole count.  The amount is now a
double and the ten call sites pass what they computed.

The computation around it was single precision too -- both source
values, the amount and their product -- where upstream is double
throughout.  The truncation points are unchanged: the cache still holds
one modulator's contribution as an Int16Array store would, and the two
recompute paths still round where they did, because upstream's do.

gs_controller_matrix_comparison 77.74 -> 135.29 dB, and it was the last
file in the corpus below 80.  The corpus is now 1 very-close and 30
near-identical.

+++ spessasynth_core/include/spessasynth/soundbank/soundbank.h
@@ -40,7 +40,7 @@ typedef struct {
 	uint16_t source_enum; /* packed ModulatorSource */
 	uint16_t amount_source_enum;
 	uint16_t dest_enum; /* SS_GeneratorType target */
-	int16_t transform_amount;
+	double transform_amount; /* fractional: the GS controller matrix sets it so */
 	uint16_t transform_type;
 	float current_value; /* runtime: computed modulator output */
 	bool is_effect_modulator;
